+++ src-docs/src/views/icon/apps.js
@@ -19,7 +19,6 @@ const iconTypes = [
   'dashboardApp',
   'devToolsApp',
   'discoverApp',
-  'emsApp',
   'filebeatApp',
   'gisApp',
   'graphApp',